requirements.txt
pytest==7.4.3
playwright==1.40.0
pytest-playwright==0.4.3
python-dotenv==1.0.0
pytest-html==4.1.1
psycopg2-binary==2.9.9